test(captions): relocate roundtrip test under _captions mirror (audit PS-204)
